+++ index.php
@@ -1,6 +1,17 @@
 <?php $awada_theme_options = awada_theme_options();
-get_header();
-get_template_part('breadcrumbs'); ?>
+get_header(); ?>
+<section class="post-wrapper-top jt-shadow clearfix">
+	<div class="container">
+		<div class="col-lg-12">
+		<?php if(class_exists('woocommerce') && is_shop()){
+			echo '<h2>'.__('Shop','awada').'</h2>';
+			}else{?>
+			<h2><?php if(is_front_page()){echo "Home";}else{single_post_title();} ?></h2>
+			<?php } ?>
+			<?php awada_breadcrumbs(); ?>
+		</div>
+	</div>
+</section><!-- end post-wrapper-top -->
 <section class="blog-wrapper">
 	<div class="container">
 		<?php